requirements.txt
streamlit>=1.31.0
supabase>=2.3.0
python-dotenv>=1.0.0
pandas>=2.1.0
plotly>=5.18.0
pydantic>=2.5.0
streamlit-option-menu>=0.3.6
